pre correction module06

+++ 42CPP-06/ex01/includes/Data.hpp
@@ -14,7 +14,7 @@ class Data{
 		Data& operator=(Data const& rhs);
 
 		int			_firstData;
-		int			_secondData;
+		float		_secondData;
 		std::string	_differentData;
 };
